🩹 [Patch]: Add retry on Install-PSResource (#27)
## Description

This pull request includes a change to the `scripts/main.ps1`
introducing a retry mechanism for the `Install-PSResource` command to
handle intermittent failures more gracefully.

Improvements to installation process:

*
[`scripts/main.ps1`](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011R23-R37):
Added retry logic with a maximum of 5 attempts and a 10-second delay
between attempts to the `Install-PSResource` command. This change helps
ensure that temporary issues do not cause the entire installation
process to fail.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas