🩹 [Patch]: Add WorkingDirectory input to action configuration and README (#31)
## Description

This pull request introduces a new input parameter to specify the
working directory for the script execution in the PSModule framework.
The changes are reflected in the documentation and the action
configuration file.

### Documentation Update

* `README.md`:
* Added a description for the new `WorkingDirectory` parameter,
specifying its default value as `${{ github.workspace }}`.

### Action Configuration

* `action.yml`:
* Introduced the `WorkingDirectory` input parameter with a description,
default value, and its inclusion in the `runs` section to pass the value
to the script.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas